+++ src/hooks/useCirculationLog.js
@@ -32,10 +32,14 @@ export const useCirculationLog = (isLoadingRightAway, queryLoadRecords, loadReco
 
     setIsLoading(true);
 
-    return queryLoadRecords(defaultSearchParams.offset, hasToCallAPI).then(recordsResponse => {
-      if (!offset) {
-        setRecordsCount(recordsResponse?.totalRecords);
+    // to get the actual totalRecords count, the limit should be set to 0
+    // Don't wait for it, because sometimes it takes twice as long as requesting the records below without a zero limit
+    queryLoadRecords(defaultSearchParams.offset, 0).then(totalRecordsResponse => {
+      setRecordsCount(totalRecordsResponse?.totalRecords);
+    });
 
+    return queryLoadRecords(defaultSearchParams.offset).then(recordsResponse => {
+      if (!offset) {
         if (recordsResponse?.totalRecords != null) {
           // eslint-disable-next-line no-unused-expressions
           resultsPaneTitleRef?.current?.focus();
@@ -47,9 +51,12 @@ export const useCirculationLog = (isLoadingRightAway, queryLoadRecords, loadReco
   }, [queryParams, isLoadingRightAway, isLoading, queryLoadRecords, defaultSearchParams.offset, loadRecordsCB]);
 
   const refreshList = useCallback(() => {
+    if (!location.search) {
+      setRecordsCount(0);
+    }
     setRecords([]);
     loadRecords(0);
-  }, [loadRecords]);
+  }, [loadRecords, location.search]);
 
   useEffect(
     () => {